Move and split tests
